requirements.txt
asgiref==3.8.1
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
dnspython==2.6.1
filelock==3.16.0
fonttools==4.53.1
fsspec==2024.9.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.7
lapx==0.5.10.post1
MarkupSafe==2.1.5
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.2
pillow==10.4.0
psutil==6.0.0
py-cpuinfo==9.0.0
pycparser==2.22
pycryptodome==3.20.0
PyJWT==2.9.0
pymongo==4.8.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
scipy==1.14.1
seaborn==0.13.2
setuptools==75.0.0
shapely==2.0.6
six==1.16.0
sqlparse==0.5.1
sympy==1.13.2
torch==2.4.1+cu124
torchaudio==2.4.1+cu124
torchvision==0.19.1+cu124
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.84
ultralytics-thop==2.0.6
urllib3==2.2.3
Werkzeug==3.0.4
